refactor(watcher): centralize ignore folders logic in utils.py
